+++ src/pl/plajer/villagedefense3/language/LanguageMigrator.java
@@ -18,6 +18,7 @@
 
 package pl.plajer.villagedefense3.language;
 
+import org.apache.commons.io.FileUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.plugin.java.JavaPlugin;
@@ -26,83 +27,103 @@ import pl.plajer.villagedefense3.handlers.ConfigurationManager;
 import pl.plajer.villagedefense3.utils.MessageUtils;
 
 import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
+/*
+  NOTE FOR CONTRIBUTORS - Please do not touch this class if you don't now how it works! You can break migrator modyfing these values!
+ */
 public class LanguageMigrator {
 
+    public static final int LANGUAGE_FILE_VERSION = 4;
     private static Main plugin = JavaPlugin.getPlugin(Main.class);
     private static List<String> migratable = Arrays.asList("bungee", "config", "kits", "language", "lobbyitems", "mysql");
 
     public static void configUpdate() {
         if(plugin.getConfig().getString("Version").equals("1")) {
             Bukkit.getConsoleSender().sendMessage(ChatColor.YELLOW + "[Village Defense] System notify >> Your config file is outdated! Updating...");
-            plugin.getConfig().set("Powerups.Enabled", true);
-            plugin.getConfig().set("Powerups.Drop-Chance", 1.0);
-            plugin.getConfig().set("Powerups.List.Map-Clean", true);
-            plugin.getConfig().set("Powerups.List.Double-Damage-For-Players.Enabled", true);
-            plugin.getConfig().set("Powerups.List.Double-Damage-For-Players.Time", 15);
-            plugin.getConfig().set("Powerups.List.Healing-For-Players.Enabled", true);
-            plugin.getConfig().set("Powerups.List.Healing-For-Players.Amplifier", 1);
-            plugin.getConfig().set("Powerups.List.Healing-For-Players.Time-Of-Healing", 15);
-            plugin.getConfig().set("Powerups.List.Golem-Raid.Enabled", true);
-            plugin.getConfig().set("Powerups.List.Golem-Raid.Golems-Amount", 3);
-            plugin.getConfig().set("Powerups.List.One-Shot-One-Kill.Enabled", true);
-            plugin.getConfig().set("Powerups.List.One-Shot-One-Kill.Time", 15);
-            plugin.getConfig().set("Version", 2);
-            plugin.saveConfig();
-            Bukkit.getConsoleSender().sendMessage(ChatColor.GREEN + "[Village Defense] System notify >> Config updated, however comments of file are lost! Please reset file if you wish to get comments again.");
+            File file = new File(plugin.getDataFolder() + "/config.yml");
+            try {
+                LanguageMigrator.removeLineFromFile(file, "# Don't modify.");
+                LanguageMigrator.removeLineFromFile(file, "Version: 1");
+                LanguageMigrator.removeLineFromFile(file, "# No way! You've reached the end! But... where's the dragon!?");
+                LanguageMigrator.addNewLines(file, "# Power ups section. If you want to have classic Village Defense game mode i recommend to disable this.\r\nPowerups:\r\n" +
+                        "  # Do you want to enable in-game power ups?\r\n  # This will make zombies to drop some power ups when they're killed\r\n" +
+                        "  # REQUIRES Holographic Displays otherwise it won't be enabled!\r\n  Enabled: true\r\n  # Modify powerup drop chance here\r\n" +
+                        "  Drop-Chance: 1.0 # 1% chance by default\r\n  # Enable or disable specific power ups here.\r\n  List:\r\n" +
+                        "    Map-Clean: true\r\n    Double-Damage-For-Players:\r\n      Enabled: true\r\n      Time: 15 # seconds\r\n" +
+                        "    Healing-For-Players:\r\n      Enabled: true\r\n      Amplifier: 1\r\n      Time-Of-Healing: 10 # seconds\r\n" +
+                        "    # Spawns X golems in village\r\n    # Owner of golems is person who picked up power up\r\n    Golem-Raid:\r\n" +
+                        "      Enabled: true\r\n      Golems-Amount: 3\r\n    # Every zombie can be killed for one hit\r\n    One-Shot-One-Kill:\r\n" +
+                        "      Enabled: true\r\n      Time: 15 # seconds\r\n\r\n" +
+                        "# Don't modify\r\nVersion: 2\r\n\r\n# No way! You've reached the end! But... where's the dragon!?");
+            } catch(IOException e) {
+                e.printStackTrace();
+                plugin.getLogger().warning("Something went horribly wrong with migration! Please contact author!");
+            }
+            Bukkit.getConsoleSender().sendMessage(ChatColor.GREEN + "[Village Defense] System notify >> Config updated, no comments were removed :)");
             return;
         }
         Bukkit.getConsoleSender().sendMessage(ChatColor.GREEN + "[Village Defense] System notify >> You're using latest config file version! Nice!");
     }
 
     public static void languageFileUpdate() {
-        if(LanguageManager.getDefaultLanguageMessage("File-Version-Do-Not-Edit").equals("0")) {
-            Bukkit.getConsoleSender().sendMessage(ChatColor.YELLOW + "[Village Defense] System notify >> Your language file is outdated! Updating...");
-            LanguageManager.getLanguageFile().set("In-Game.Spectator.Target-Player-Health", "&cHealth: &7%health%");
-            LanguageManager.getLanguageFile().set("Scoreboard.Footer", "&ewww.spigotmc.org");
-            LanguageManager.getLanguageFile().set("File-Version-Do-Not-Edit", 1);
-            LanguageManager.saveLanguageFile();
-            Bukkit.getConsoleSender().sendMessage(ChatColor.GREEN + "[Village Defense] System notify >> Language file updated! Nice!");
+        if(LanguageManager.getDefaultLanguageMessage("File-Version-Do-Not-Edit").equals(String.valueOf(LANGUAGE_FILE_VERSION))) return;
+        Bukkit.getConsoleSender().sendMessage(ChatColor.YELLOW + "[Village Defense] System notify >> Your language file is outdated! Updating...");
+
+        int version = Integer.valueOf(LanguageManager.getDefaultLanguageMessage("File-Version-Do-Not-Edit"));
+        LanguageMigrator.updateLanguageVersionControl(version);
+
+        File file = new File(plugin.getDataFolder() + "/language.yml");
+
+        //todo simplify
+        switch(version) {
+            case 0:
+                LanguageMigrator.insertAfterLine(file, "Spectator-Menu-Name", "    Target-Player-Health: \"&cHealth: &7%health%\"");
+                LanguageMigrator.insertAfterLine(file, "Scoreboard:", "  Footer: \"&ewww.spigotmc.org\"");
+                LanguageMigrator.addNewLines(file, "\r\nPowerups:\r\n  Map-Clean-Powerup:\r\n    Name: \"&e&lMAP CLEANER\"\r\n    # Used as sub title description\r\n" +
+                        "    Description: \"&7Map has been cleaned!\"\r\n  Double-Damage-Powerup:\r\n    Name: \"&c&lDOUBLE DAMAGE\"\r\n    Description: \"&7Double damage for %time% seconds!\"\r\n" +
+                        "  Healing-Powerup:\r\n    Name: \"&6&lREJUVENATION\"\r\n    Description: \"&7Healing for %time% seconds!\"\r\n  Golem-Raid-Powerup:\r\n    Name: \"&a&lIRONBOUND RAID\"\r\n" +
+                        "    Description: \"&7Golems have invaded this village!\"\r\n  One-Shot-One-Kill-Powerup:\r\n    Name: \"&b&lFRENZY\"\r\n    Description: \"&7Every zombie for 1 hit!\"\r\n" +
+                        "  Powerup-Ended-Title-Message: \"&4Powerup %powerup% has ended!\"\r\n");
+                LanguageMigrator.insertAfterLine(file, "Commands:", "  Statistics:");
+                LanguageMigrator.insertAfterLine(file, "Statistics:", "    Format: \"&e#%position% %name% - %value% &7%statistic%\"");
+                LanguageMigrator.insertAfterLine(file, "Statistics:", "    Header: \"&8&m-------------------[&6 Top 10 &8&m]-------------------\"");
+                LanguageMigrator.insertAfterLine(file, "Statistics:", "    Invalid-Name: \"&cName of statistic is invalid! Type: kills, deaths, games_played, highest_wave, level\"");
+                LanguageMigrator.insertAfterLine(file, "Statistics:", "    Type-Name: \"&cPlease type statistic name to view!\"");
+                LanguageMigrator.insertAfterLine(file, "Commands:", "  Did-You-Mean: \"&6Did you mean &7/%command%&6?\"");
+                break;
+            case 1:
+                LanguageMigrator.addNewLines(file, "\r\nPowerups:\r\n  Map-Clean-Powerup:\r\n    Name: \"&e&lMAP CLEANER\"\r\n    # Used as sub title description\r\n" +
+                        "    Description: \"&7Map has been cleaned!\"\r\n  Double-Damage-Powerup:\r\n    Name: \"&c&lDOUBLE DAMAGE\"\r\n    Description: \"&7Double damage for %time% seconds!\"\r\n" +
+                        "  Healing-Powerup:\r\n    Name: \"&6&lREJUVENATION\"\r\n    Description: \"&7Healing for %time% seconds!\"\r\n  Golem-Raid-Powerup:\r\n    Name: \"&a&lIRONBOUND RAID\"\r\n" +
+                        "    Description: \"&7Golems have invaded this village!\"\r\n  One-Shot-One-Kill-Powerup:\r\n    Name: \"&b&lFRENZY\"\r\n    Description: \"&7Every zombie for 1 hit!\"\r\n" +
+                        "  Powerup-Ended-Title-Message: \"&4Powerup %powerup% has ended!\"\r\n");
+                LanguageMigrator.insertAfterLine(file, "Commands:", "  Statistics:");
+                LanguageMigrator.insertAfterLine(file, "Statistics:", "    Format: \"&e#%position% %name% - %value% &7%statistic%\"");
+                LanguageMigrator.insertAfterLine(file, "Statistics:", "    Header: \"&8&m-------------------[&6 Top 10 &8&m]-------------------\"");
+                LanguageMigrator.insertAfterLine(file, "Statistics:", "    Invalid-Name: \"&cName of statistic is invalid! Type: kills, deaths, games_played, highest_wave, level\"");
+                LanguageMigrator.insertAfterLine(file, "Statistics:", "    Type-Name: \"&cPlease type statistic name to view!\"");
+                LanguageMigrator.insertAfterLine(file, "Commands:", "  Did-You-Mean: \"&6Did you mean &7/%command%&6?\"");
+                break;
+            case 2:
+                LanguageMigrator.insertAfterLine(file, "Commands:", "  Statistics:");
+                LanguageMigrator.insertAfterLine(file, "Statistics:", "    Format: \"&e#%position% %name% - %value% &7%statistic%\"");
+                LanguageMigrator.insertAfterLine(file, "Statistics:", "    Header: \"&8&m-------------------[&6 Top 10 &8&m]-------------------\"");
+                LanguageMigrator.insertAfterLine(file, "Statistics:", "    Invalid-Name: \"&cName of statistic is invalid! Type: kills, deaths, games_played, highest_wave, level\"");
+                LanguageMigrator.insertAfterLine(file, "Statistics:", "    Type-Name: \"&cPlease type statistic name to view!\"");
+                LanguageMigrator.insertAfterLine(file, "Commands:", "  Did-You-Mean: \"&6Did you mean &7/%command%&6?\"");
+                break;
+            case 3:
+                LanguageMigrator.insertAfterLine(file, "Commands:", "  Did-You-Mean: \"&6Did you mean &7/%command%&6?\"");
+                break;
         }
-        if(LanguageManager.getDefaultLanguageMessage("File-Version-Do-Not-Edit").equals("1")) {
-            Bukkit.getConsoleSender().sendMessage(ChatColor.YELLOW + "[Village Defense] System notify >> Your language file is outdated! Updating...");
-            LanguageManager.getLanguageFile().set("Powerups.Map-Clean-Powerup.Name", "&e&lMAP CLEANER");
-            LanguageManager.getLanguageFile().set("Powerups.Map-Clean-Powerup.Description", "&7Map has been cleaned!");
-            LanguageManager.getLanguageFile().set("Powerups.Double-Damage-Powerup.Name", "&c&lDOUBLE DAMAGE");
-            LanguageManager.getLanguageFile().set("Powerups.Double-Damage-Powerup.Description", "&7Double damage for %time% seconds!");
-            LanguageManager.getLanguageFile().set("Powerups.Healing-Powerup.Name", "&6&lREJUVENATION");
-            LanguageManager.getLanguageFile().set("Powerups.Healing-Powerup.Description", "&7Healing for %time% seconds!");
-            LanguageManager.getLanguageFile().set("Powerups.Golem-Raid-Powerup.Name", "&a&lIRONBOUND RAID");
-            LanguageManager.getLanguageFile().set("Powerups.Golem-Raid-Powerup.Description", "&7Golems have invaded this village!");
-            LanguageManager.getLanguageFile().set("Powerups.One-Shot-One-Kill-Powerup.Name", "&b&lFRENZY");
-            LanguageManager.getLanguageFile().set("Powerups.One-Shot-One-Kill-Powerup.Description", "&7Every zombie for 1 hit!");
-            LanguageManager.getLanguageFile().set("Powerups.Powerup-Ended-Title-Message", "&4Powerup %powerup% has ended!");
-            LanguageManager.getLanguageFile().set("File-Version-Do-Not-Edit", 2);
-            LanguageManager.saveLanguageFile();
-            Bukkit.getConsoleSender().sendMessage(ChatColor.GREEN + "[Village Defense] System notify >> Language file updated! Nice!");
-            return;
-        }
-        if(LanguageManager.getDefaultLanguageMessage("File-Version-Do-Not-Edit").equals("2")) {
-            Bukkit.getConsoleSender().sendMessage(ChatColor.YELLOW + "[Village Defense] System notify >> Your language file is outdated! Updating...");
-            LanguageManager.getLanguageFile().set("Commands.Statistics.Type-Name", "&cPlease type statistic name to view!");
-            LanguageManager.getLanguageFile().set("Commands.Statistics.Invalid-Name", "&cName of statistic is invalid! Type: kills, deaths, games_played, highest_wave, level");
-            LanguageManager.getLanguageFile().set("Commands.Statistics.Header", "&8&m-------------------[&6 Top 10 &8&m]-------------------");
-            LanguageManager.getLanguageFile().set("Commands.Statistics.Format", "&e#%position% %name% - %value% &7%statistic%");
-            LanguageManager.getLanguageFile().set("File-Version-Do-Not-Edit", 3);
-            LanguageManager.saveLanguageFile();
-            Bukkit.getConsoleSender().sendMessage(ChatColor.GREEN + "[Village Defense] System notify >> Language file updated! Nice!");
-            return;
-        }
-        if(LanguageManager.getDefaultLanguageMessage("File-Version-Do-Not-Edit").equals("3")) {
-            Bukkit.getConsoleSender().sendMessage(ChatColor.YELLOW + "[Village Defense] System notify >> Your language file is outdated! Updating...");
-            LanguageManager.getLanguageFile().set("Commands.Did-You-Mean", "&6Did you mean &7/%command%&6?");
-            LanguageManager.getLanguageFile().set("File-Version-Do-Not-Edit", 4);
-            LanguageManager.saveLanguageFile();
-            Bukkit.getConsoleSender().sendMessage(ChatColor.GREEN + "[Village Defense] System notify >> Language file updated! Nice!");
-            return;
-        }
+        Bukkit.getConsoleSender().sendMessage(ChatColor.GREEN + "[Village Defense] System notify >> Language file updated! Nice!");
         Bukkit.getConsoleSender().sendMessage(ChatColor.GREEN + "[Village Defense] System notify >> You're using latest language file version! Nice!");
     }
 
@@ -117,6 +138,53 @@ public class LanguageMigrator {
             }
         }
         Bukkit.getConsoleSender().sendMessage(ChatColor.GREEN + "Done! Enabling VD3...");
+    }
+
+    private static void removeLineFromFile(File file, String lineToRemove) throws IOException {
+        List<String> lines = FileUtils.readLines(file);
+        List<String> updatedLines = lines.stream().filter(s -> !s.contains(lineToRemove)).collect(Collectors.toList());
+        FileUtils.writeLines(file, updatedLines, false);
+    }
+
+    private static void insertAfterLine(File file, String search, String text) {
+        try {
+            int i = 1;
+            List<String> lines = Files.readAllLines(file.toPath(), StandardCharsets.UTF_8);
+            for(String line : lines) {
+                if(line.contains(search)) {
+                    lines.add(i, text);
+                    Files.write(file.toPath(), lines, StandardCharsets.UTF_8);
+                    break;
+                }
+                i++;
+            }
+        } catch(IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    private static void updateLanguageVersionControl(int oldVersion) {
+        try {
+            File file = new File(plugin.getDataFolder() + "/language.yml");
+            LanguageMigrator.removeLineFromFile(file, "# Don't edit it. But who's stopping you? It's your server!");
+            LanguageMigrator.removeLineFromFile(file, "# Really, don't edit ;p");
+            LanguageMigrator.removeLineFromFile(file, "File-Version-Do-Not-Edit: " + oldVersion);
+            LanguageMigrator.addNewLines(file, "# Don't edit it. But who's stopping you? It's your server!\n# Really, don't edit ;p\nFile-Version-Do-Not-Edit: " + LANGUAGE_FILE_VERSION);
+        } catch(IOException e) {
+            e.printStackTrace();
+            plugin.getLogger().warning("Something went horribly wrong with migration! Please contact author!");
+        }
+    }
+
+    private static void addNewLines(File file, String newLines) {
+        try {
+            FileWriter fw = new FileWriter(file.getPath(), true);
+            fw.write(newLines);
+            fw.close();
+        } catch(IOException e) {
+            e.printStackTrace();
+            plugin.getLogger().warning("Something went horribly wrong with migration! Please contact author!");
+        }
     }
 
 }